// src/models/PolicyInfo.ts
export interface PolicyInfo {
    policyId:                     string;
    folio:                        string;
    folioOT:                      null;
    insuranceId:                  string;
    insuranceCompany:             string;
    branchId:                     string;
    subBranchId:                  string;
    noPolicy:                     string;
    endorsement:                  number;
    agentNumber:                  string;
    renewablePolicy:              boolean;
    issuanceDate:                 Date;
    startValidity:                Date;
    endValidity:                  Date;
    currency:                     string;
    paymentMethod:                string;
    collectionMethod:             string;
    clientId:                     string;
    clientName:                   string;
    rfc:                          string;
    street:                       string;
    neighborhood:                 string;
    state:                        number;
    municipality:                 number;
    locality:                     number;
    zip:                          number;
    country:                      string;
    salesPerson:                  string;
    coveragePackageFolio:         string;
    coveragePackageByPolicyFolio: string;
    groups:                       string;
    comments:                     string;
    netPremium:                   number;
    settingOne:                   number;
    settingTwo:                   number;
    rights:                       number;
    financing:                    number;
    iva:                          number;
    totalPremium:                 number;
    additionalCharge:             number;
    subtotal:                     number;
    receipts:                     number;
    commissionPercentage:         number;
    policyStatus:                 string;
    objectStatusId:               number;
    createdAt:                    Date;
    createdBy:                    string;
    updatedAt:                    Date;
    updatedBy:                    string;
    concept:                      string;
    health:                       number;
}

export default PolicyInfo